Split JsonTokenTest into two files
